+++ packages/flutter_bloc/lib/src/bloc_builder.dart
@@ -4,7 +4,7 @@ import 'package:flutter_cubit/flutter_cubit.dart';
 
 import '../flutter_bloc.dart';
 
-/// {@template blocbuilder}
+/// {@template bloc_builder}
 /// [BlocBuilder] handles building a widget in response to new [states].
 /// [BlocBuilder] is analogous to [StreamBuilder] but has simplified API to
 /// reduce the amount of boilerplate code needed as well as [bloc]-specific
@@ -36,20 +36,19 @@ import '../flutter_bloc.dart';
 /// )
 /// ```
 ///
-/// An optional [condition] can be implemented for more granular control over
+/// An optional [buildWhen] can be implemented for more granular control over
 /// how often [BlocBuilder] rebuilds.
-/// The [condition] function will be invoked on each [bloc] [state] change.
-/// The [condition] takes the previous [state] and current [state] and must
+/// [buildWhen] will be invoked on each [cubit] `state` change.
+/// [buildWhen] takes the previous `state` and current `state` and must
 /// return a [bool] which determines whether or not the [builder] function will
 /// be invoked.
-/// The previous [state] will be initialized to the [state] of the [bloc] when
+/// The previous `state` will be initialized to the `state` of the [bloc] when
 /// the [BlocBuilder] is initialized.
-/// [condition] is optional and if it isn't implemented, it will default to
-/// `true`.
+/// [buildWhen] is optional and if omitted, it will default to `true`.
 ///
 /// ```dart
 /// BlocBuilder<BlocA, BlocAState>(
-///   condition: (previous, current) {
+///   buildWhen: (previous, current) {
 ///     // return true/false to determine whether or not
 ///     // to rebuild the widget with state
 ///   },
@@ -60,12 +59,12 @@ import '../flutter_bloc.dart';
 /// ```
 /// {@endtemplate}
 class BlocBuilder<B extends Bloc<dynamic, S>, S> extends CubitBuilder<B, S> {
-  /// {@macro blocbuilder}
+  /// {@macro bloc_builder}
   const BlocBuilder({
     Key key,
     @required CubitWidgetBuilder<S> builder,
     B bloc,
-    CubitBuilderCondition<S> condition,
+    CubitBuilderCondition<S> buildWhen,
   })  : assert(builder != null),
-        super(key: key, builder: builder, cubit: bloc, condition: condition);
+        super(key: key, builder: builder, cubit: bloc, buildWhen: buildWhen);
 }